Factor out filter stage into it's own file.
