+++ pkg/query/plan/pipeline.go
@@ -155,56 +155,3 @@ func (c *CollectStage) Add(entries []WrappedEntry) {
 		c.Output <- entry
 	}
 }
-
-type FilterStage struct {
-	next  Stage
-	root  *ast.DataFunctionNode
-	input chan []WrappedEntry
-	once  sync.Once
-}
-
-func MakeFilterStage(node *ast.DataFunctionNode) *FilterStage {
-	var f FilterStage
-
-	f.input = make(chan []WrappedEntry)
-	f.root = node
-	return &f
-}
-
-func (f *FilterStage) Chain(next Stage) {
-	f.next = next
-}
-
-func (f *FilterStage) Next() Stage {
-	return f.next
-}
-
-func (f *FilterStage) Add(entries []WrappedEntry) {
-	f.input <- entries
-}
-
-func (f *FilterStage) Finish() {
-	f.once.Do(func() {
-		close(f.input)
-	})
-}
-
-func (f *FilterStage) Execute() {
-	for entries := range f.input {
-		symbols := make(SymbolMap)
-
-		for idx, arg := range f.root.Arguments {
-			symbols[arg.Value()] = entries[idx].Value()
-		}
-
-		fn := MakeFunction(symbols)
-		ast.Walk(&fn, f.root)
-
-		allowed := types.BooleanVal(fn.Result[0])
-
-		if allowed {
-			f.Next().Add(entries)
-		}
-	}
-	f.Next().Finish()
-}
